Admin Roles & Permission Module

+++ app/Http/Controllers/admin/CategoryController.php
@@ -10,7 +10,7 @@ use Illuminate\Http\Request;
 class CategoryController extends Controller
 {
     public function category(){
-
+        abort_unless($this->checkPermission('View Category'), 403);
         $name = isset($_REQUEST['name']) ? trim($_REQUEST['name']) : "";
         $status = isset($_REQUEST['status']) ? trim($_REQUEST['status']) : "";
 
@@ -29,13 +29,15 @@ class CategoryController extends Controller
         return view('admin.category.index')->with('data',$data);
     }
 
-    public function create_category(){      
+    public function create_category(){
+        abort_unless($this->checkPermission('Create Category'), 403);
         $data['pageTitle'] = "Add Category";
         $data['action'] = "Add";
         return view('admin.category.category_action')->with('data',$data);
     }
     
     public function store_category(Request $request){
+        abort_unless($this->checkPermission('Create Category'), 403);
         $validated = $request->validate([
             'category_name' => 'required|max:255',
             'status' => 'required',
@@ -48,6 +50,7 @@ class CategoryController extends Controller
     }
 
     public function edit_category($id){
+        abort_unless($this->checkPermission('Edit Category'), 403);
         $dataCategory = Category::find($id);
         $data['pageTitle'] = "Edit Category";
         $data['action'] = "Edit";
@@ -56,6 +59,7 @@ class CategoryController extends Controller
     }
 
     public function update_category($id,Request $request){
+        abort_unless($this->checkPermission('Edit Category'), 403);
         $validated = $request->validate([
             'category_name' => 'required|max:255',
             'status' => 'required',
@@ -68,6 +72,7 @@ class CategoryController extends Controller
     }
 
     public function destroy_category(Request $request){
+        abort_unless($this->checkPermission('Delete Category'), 403);
         $category = Category::find($request->dataId);
         $category->children()->delete();
         $category->products()->delete();
@@ -76,6 +81,7 @@ class CategoryController extends Controller
     }
     
     public function subcategory(){
+        abort_unless($this->checkPermission('View SubCategory'), 403);
         $name = isset($_REQUEST['name']) ? trim($_REQUEST['name']) : "";
         $status = isset($_REQUEST['status']) ? trim($_REQUEST['status']) : "";
 
@@ -94,12 +100,14 @@ class CategoryController extends Controller
         return view('admin.category.subcategory')->with('data',$data);
     }
     public function create_subcategory(){
+        abort_unless($this->checkPermission('Create SubCategory'), 403);
         $data['pageTitle'] = "Add SubCategory";
         $data['action'] = "Add";
         $data['pageData']['category'] = Category::where([['parent_id','=',0]])->orderBy('id', 'desc')->get();
         return view('admin.category.subcategory_action')->with('data',$data);
     }
     public function store_subcategory(Request $request){
+        abort_unless($this->checkPermission('Create SubCategory'), 403);
         $validated = $request->validate([
             'parent_id' => 'required|integer',
             'category_name' => 'required|max:255',
@@ -111,6 +119,7 @@ class CategoryController extends Controller
         return redirect()->route('admin.subcategory')->with('success','Data Added Successfuly');
     }
     public function edit_subcategory($id,Request $request){
+        abort_unless($this->checkPermission('Edit SubCategory'), 403);
         $dataCategory = Category::find($id);
         $data['pageTitle'] = "Edit SubCategory";
         $data['action'] = "Edit";
@@ -119,6 +128,7 @@ class CategoryController extends Controller
         return view('admin.category.subcategory_action')->with('data',$data);
     }
     public function update_subcategory($id,Request $request){
+        abort_unless($this->checkPermission('Edit SubCategory'), 403);
         $validated = $request->validate([
             'parent_id' => 'required|integer',
             'category_name' => 'required|max:255',
@@ -130,12 +140,14 @@ class CategoryController extends Controller
         return redirect()->route('admin.subcategory')->with('success','Data Updated Successfuly');
     }
     public function destroy_subcategory(Request $request){
+        abort_unless($this->checkPermission('Delete SubCategory'), 403);
         $category = Category::find($request->dataId);
         $category->delete();
         echo 1;
     }
 
     public function get_ajax_category(Request $request){
+        abort_unless($this->checkPermission('View Category'), 403);
         $selectedId = $request->input('selectedId');
         $catData = Builder::getCategoryData();
         if( $catData->count() > 0){
@@ -148,6 +160,7 @@ class CategoryController extends Controller
         return response()->json(['success' => true, 'subCatData' => $CatData]);
     }
     public function get_ajax_subcategory(Request $request){
+        abort_unless($this->checkPermission('View SubCategory'), 403);
         $categoryId = $request->input('categoryId');
         $selectedId = $request->input('selectedId');        
         $subCat = Builder::getSubCategoryData('*',['status'=>'Active','parent_id'=>$categoryId]);
